Allow both strings and booleans for i18n component ‘tag’ prop
This change widens the type for the i18n component's ‘tag’ prop to
also allow booleans.

At runtime ‘:tag=false’ works by rendering no tag, but TypeScript
complains that passing a boolean (instead of a string) is not allowed.

Note: only the literal ‘false’ is intended to be used, but using
literals/const in the type definition does not work, so this type is
slightly wider than it should be. The alternative is to add ts-ignores
or ‘as any’ casts (which eslint may then complain about) everywhere
this feature is used, which is a worse situation.